+++ src/features/tasks/components/CreateTask.tsx
@@ -1,22 +1,28 @@
 import React from 'react';
 
+// components
 import { TaskForm } from './TaskForm';
 
-import { TaskIncompleteType } from '../../types';
-import { ADD_TASK } from '../../store/actions';
-import { useStore } from '../../store/useStore';
+// store
+import { useStore } from '@/store/useStore';
+
+// types
+import type { TaskIncompleteType } from '@/types';
 
 type Props = {
   onClose: () => void;
   onDiscard: () => void;
 };
 
+/**
+ * Handles the logic of creating tasks, and composes TaskForm to render the form ui.
+ */
 export const CreateTask: React.FC<Props> = ({ onClose, onDiscard }) => {
   const { dispatch } = useStore();
 
   const handleSubmit = (data: TaskIncompleteType) => {
     dispatch({
-      type: ADD_TASK,
+      type: 'ADD_TASK',
       payload: { ...data },
     });
     onClose();
