+++ src/features/tasks/components/__tests__/TaskForm.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '@testing-library/jest-dom';
-import { render, screen, cleanup, waitFor } from '../../../utils/test-utils';
+import { render, screen, cleanup, waitFor } from '@/utils/test-utils';
 import userEvent from '@testing-library/user-event';
 
 // component
